# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Criasol - backend",
        "description": "api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "operationId": "f92ff81e405f6f6224ed353490975ced",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "musk@email.com",
                                    "password": "password1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Logout",
                "operationId": "ac1c8a886ecd17e108b62c43687439ed",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Register",
                "operationId": "e7e369b43858339ddf6d345e5dfdf7a1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Musk",
                                    "email": "musk@email.com",
                                    "password": "password1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/example": {
            "get": {
                "tags": [
                    "example"
                ],
                "summary": "Example",
                "description": "Get Exemplo",
                "operationId": "2bea81dd24e5d9c2e49760b4310620e1",
                "responses": {
                    "default": {
                        "description": "List Items"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}